+++ CHANGELOG.md
@@ -56,6 +56,8 @@ Date format: (YYYY-MM-DD)
 * The spawning and despawning of shopkeepers is handled more consistently now. For instance, if shopkeepers are marked as dirty during the spawning of their shop objects we would previously trigger a delayed save. We do the same when they are despawned now.
 * Debug: Several tasks no longer use lambdas but named classes. This makes it easier to identify them in timing reports.
 * Debug: A few debug messages are now printed before their denoted action takes place (eg. before the shop object is teleported or respawned).
+* Debug: The debug command can toggle debug options now ('/shopkeeper debug [option]').
+* Fixed: Optional arguments delegate to the wrapped argument for more detailed error messages now.
 * Performance: Living shopkeepers no longer tick Minecraft's floating behavior. This has no effect since the shopkeeper entities have no AI.
 
 API:  
@@ -83,6 +85,7 @@ Internal:
 * Added AbstractShopkeeper#onChunkActivation() and #onChunkDeactivation().
 * Instead of passing the chunk coordinates around whenever a shopkeeper is registered, unregistered, or moved, each shopkeeper remembers the chunk now by which we previously stored it.
 * ShopkeeperRegistry#onShopkeeperMove has been renamed to #onShopkeeperMoved.
+* The FixedValuesArgument is less restrictive with the types of Map values it accepts.
 * Minor code formatting and method renaming.
 
 Migration notes: